+++ test/reviver.spec.ts
@@ -6,6 +6,13 @@ import {
   isEither,
   isOption,
   isContainer,
+  LIST_OBJECT_TYPE,
+  IDLE_OBJECT_TYPE,
+  TUPLE_OBJECT_TYPE,
+  CONTAINER_OBJECT_TYPE,
+  OPTION_SOME_OBJECT_TYPE,
+  EITHER_LEFT_OBJECT_TYPE,
+  EITHER_RIGHT_OBJECT_TYPE,
 } from '../src';
 
 describe('reviver', () => {
@@ -15,39 +22,39 @@ describe('reviver', () => {
   });
 
   test('should create Option instance', () => {
-    const json = '{"type":"Some","value":1}';
+    const json = `{"type":"${OPTION_SOME_OBJECT_TYPE}","value":1}`;
     expect(isOption(JSON.parse(json, reviver))).toBe(true);
   });
 
   test('should create List instance', () => {
-    const json = '{"type":"List","value":[1]}';
+    const json = `{"type":"${LIST_OBJECT_TYPE}","value":[1]}`;
     expect(isList(JSON.parse(json, reviver))).toBe(true);
   });
 
   test('should create Either instance with Right state', () => {
-    const json = '{"type":"Right","value":1}';
+    const json = `{"type":"${EITHER_RIGHT_OBJECT_TYPE}","value":1}`;
     expect(isEither(JSON.parse(json, reviver))).toBe(true);
     expect(JSON.parse(json, reviver).isRight()).toBe(true);
   });
 
   test('should create Either instance with Left state', () => {
-    const json = '{"type":"Left","value":"Message!"}';
+    const json = `{"type":"${EITHER_LEFT_OBJECT_TYPE}","value":"Message!"}`;
     expect(isEither(JSON.parse(json, reviver))).toBe(true);
     expect(JSON.parse(json, reviver).isLeft()).toBe(true);
   });
 
   test('should create Container instance', () => {
-    const json = '{"type":"Container","value":5}';
+    const json = `{"type":"${CONTAINER_OBJECT_TYPE}","value":5}`;
     expect(isContainer(JSON.parse(json, reviver))).toBe(true);
   });
 
   test('should create Tuple instance', () => {
-    const json = '{"type":"Tuple","value":[5]}';
+    const json = `{"type":"${TUPLE_OBJECT_TYPE}","value":[5]}`;
     expect(isTuple(JSON.parse(json, reviver))).toBe(true);
   });
 
   test('should create an Idle instance', () => {
-    const json = '{"type":"Idle","value":7}';
+    const json = `{"type":"${IDLE_OBJECT_TYPE}","value":7}`;
     expect(isIdle(JSON.parse(json, reviver))).toBe(true);
   });
 });